requirements.txt
duckdb==0.10.3
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
pydeck==0.9.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
streamlit==1.36.0
